Solution for question 2 added
